no sudo
I think it may be best to take sudo out, unless you know how to fix this:
if I `sudo vim a /etc/file.conf` that will cause ntfy to send the notification--not really desired. however, If there's a very long sudo process running, like 'sudo make buildworld' on freebsd, when it's completed, ntfy won't send the notification because it's marked to ignore sudo. 
So is there an elegant way to ignore `sudo vim, tmux, emacs`, etc but watch long running processes that also run as sudo?