Avoid tainting the default config
When running tests, depending on the order they are discovered, the
default_config can become tainted before the test_default_config test is
run. This causes this test to fail.

Admittedly its likely never an issue when using ntfy as a script. But
its in general a good practice not to direcly use mutable objects as
default values and less flaky tests are easier to handle.